🐛 Fix curse adding unknown mod loader for MC version
